# swagger.yaml
openapi: 3.0.0
info:
  title: Brainrot Service API
  description: API for managing brainrot entities
  version: 1.0.0
servers:
  - url: http://localhost:4001/api/v1
    description: Local development server
#  - url: https://brainrot-service:4001/api/v1
#    description: Production server
paths:
  /brainrot:
    get:
      summary: Get all brainrots
      tags:
        - Brainrot
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Brainrot'
        '500':
          description: Internal server error
    post:
      summary: Create a new brainrot (Sigma only)
      tags:
        - Brainrot
        - Sigma
      security:
        - SigmaAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrainrotInput'
      responses:
        '201':
          description: Brainrot created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brainrot'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Sigma access required
        '500':
          description: Internal server error
  /brainrot/{id}:
    get:
      summary: Get a specific brainrot by ID
      tags:
        - Brainrot
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the brainrot to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brainrot'
        '404':
          description: Brainrot not found
        '500':
          description: Internal server error
    put:
      summary: Update a brainrot (Sigma only)
      tags:
        - Brainrot
        - Sigma
      security:
        - SigmaAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the brainrot to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrainrotInput'
      responses:
        '200':
          description: Brainrot updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brainrot'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Sigma access required
        '404':
          description: Brainrot not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a brainrot (Sigma only)
      tags:
        - Brainrot
        - Sigma
      security:
        - SigmaAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the brainrot to delete
          schema:
            type: string
      responses:
        '204':
          description: Brainrot deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Sigma access required
        '404':
          description: Brainrot not found
        '500':
          description: Internal server error
components:
  schemas:
    Brainrot:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the brainrot
        name:
          type: string
          description: Name of the brainrot
        image:
          type: string
          description: Image identifier or path
        baseHP:
          type: integer
          description: Base health points
        baseAttack:
          type: integer
          description: Base attack points
        isBoss:
          type: boolean
          description: Whether this brainrot is a boss
      required:
        - id
        - name
        - baseHP
        - baseAttack
        - isBoss
    BrainrotInput:
      type: object
      properties:
        name:
          type: string
          description: Name of the brainrot
        image:
          type: string
          description: Image identifier or path
        baseHP:
          type: integer
          description: Base health points
          minimum: 0
        baseAttack:
          type: integer
          description: Base attack points
          minimum: 0
        isBoss:
          type: boolean
          description: Whether this brainrot is a boss
          default: false
      required:
        - name
        - baseHP
        - baseAttack
  securitySchemes:
    SigmaAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Sigma JWT token required for sigma operations